Merge with upstream 1.0.15

+++ src/snark/libsnark/common/utils.cpp
@@ -41,7 +41,7 @@ size_t bitreverse(size_t n, const size_t l)
     return r;
 }
 
-bit_vector int_list_to_bits(const std::initializer_list<unsigned long long> &l, const size_t wordsize)
+bit_vector int_list_to_bits(const std::initializer_list<unsigned long> &l, const size_t wordsize)
 {
     bit_vector res(wordsize*l.size());
     for (size_t i = 0; i < l.size(); ++i)
